recreated upcoming anniversaries endpoint to get all anniversaries
